CloudFunctions/qa_pipeline/requirements.txt
functions-framework
google-cloud-firestore
google-cloud-aiplatform[adk,agent_engines]
requests
pytz
